requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic-settings
python-dotenv
pillow
aiofiles
pytest
pytest-asyncio
alembic
apscheduler 